+++ test/state_extended_test.dart
@@ -5,4 +5,4 @@
 // Calls the example app's very own test
 import '../example/test/widget_test.dart';
 
-void main() => testApp();
+void main() => testExampleApp();